chore:去掉sv模式的调整图片大小
