+++ src/ts/preview/index.ts
@@ -15,7 +15,7 @@ import {getEventName} from "../util/compatibility";
 import {hasClosestByClassName, hasClosestByMatchTag} from "../util/hasClosest";
 import {hasClosestByTag} from "../util/hasClosestByHeadings";
 import {setSelectionFocus} from "../util/selection";
-import {previewImage, resizeImage} from "./image";
+import {previewImage} from "./image";
 import { imageRender } from "../markdown/imageRender";
 
 export class Preview {
@@ -63,20 +63,13 @@ export class Preview {
                 return;
             }
             if (event.target.tagName === "IMG") {
-                if(vditor.options.image.isResize) {
-                    resizeImage(event.target as HTMLImageElement, previewElement as HTMLDivElement, vditor as IVditor)
-                }
-            }
-        });
-        previewElement.addEventListener("dblclick", (event: MouseEvent & { target: HTMLElement }) => {
-            if (event.target.tagName === "IMG") {
                 if (vditor.options.image.preview) {
                     vditor.options.image.preview(event.target)
                 } else if (vditor.options.image.isPreview) {
                     previewImage(event.target as HTMLImageElement, vditor.options.lang, vditor.options.theme);
                 }
             }
-        })
+        });
 
         const actions = vditor.options.preview.actions;
         const actionElement = document.createElement("div");
